requirements.txt
python-dotenv
hikari
hikari-lightbulb
tk
markovify
nltk
pillow
requests
tkhtmlview
markdown
BeautifulSoup4